fix(nn): dispatch moe_gemm_gguf on the device, not the feature set
The ROCm implementation was gated `all(not(feature = "cuda"), feature
= "rocm")`, so a build with both features compiled only the CUDA one.
Cargo features are additive and nothing forbids the combination -- one
dependent asking for each is enough -- and in that build `mod rocm`
and its tests vanish silently while a ROCm tensor falls into the CUDA
path and fails at `as_cuda_device()`.

`candle-core` already answers this by matching the storage at runtime
with both arms compiled in, e.g. `QStorage::device_ptr`. `moe.rs` was
the only place choosing the backend at compile time. It now matches on
`input.device()` under `any(cuda, rocm)`.

The CUDA entry points move to `moe/cuda.rs`, mirroring `moe/rocm.rs`,
which keeps `moe.rs` a dispatcher and all three files under the
400-line cap. The move is verbatim; only the `quantized::{self, ..}`
import narrows, `self` having been unused there.

`moe_gemm` is untouched: it has no ROCm implementation either way, so
its cuda/not-cuda split already covers every build.
